refactor(consensus): move staking to folder and use kebab-case; update exports/imports (#462)
