+++ packages/auto-consensus/src/index.ts
@@ -9,7 +9,7 @@ export * from './position'
 export * from './remark'
 export * from './staking'
 export * from './transfer'
-export * from './staking.withdrawal.helpers'
 
 export * from './types'
 export * from './utils'
+
