+++ packages/auto-consensus/src/staking/withdrawal-helpers.ts
@@ -1,8 +1,8 @@
 import type { ApiPromise } from '@autonomys/auto-utils'
-import { nominatorPosition } from './position'
+import { nominatorPosition } from '../position'
 import { withdrawStake } from './staking'
-import type { StringNumberOrBigInt } from './types/staking'
-import { parseString } from './utils/parse'
+import type { StringNumberOrBigInt } from '../types/staking'
+import { parseString } from '../utils/parse'
 
 export type WithdrawStakeAllParams = {
   api: ApiPromise
@@ -30,10 +30,6 @@ const clampPercent = (percent: bigint): bigint => {
   return percent
 }
 
-/**
- * Withdraw all stake for an account on an operator by converting all shares into a withdrawal.
- * Returns a submittable extrinsic.
- */
 export const withdrawStakeAll = async (params: WithdrawStakeAllParams) => {
   const { api, operatorId, account } = params
   const position = await nominatorPosition(api, operatorId, account)
@@ -44,9 +40,6 @@ export const withdrawStakeAll = async (params: WithdrawStakeAllParams) => {
   return withdrawStake({ api, operatorId, shares: totalShares })
 }
 
-/**
- * Withdraw a percentage of the current stake (rounded down in shares). Returns a submittable extrinsic.
- */
 export const withdrawStakeByPercent = async (params: WithdrawStakeByPercentParams) => {
   const { api, operatorId, account } = params
   const rawPercent = BigInt(parseString(params.percent))
@@ -57,7 +50,6 @@ export const withdrawStakeByPercent = async (params: WithdrawStakeByPercentParam
 
   if (totalShares === BigInt(0)) throw new Error('No shares to withdraw for the given account')
 
-  // shares = floor(totalShares * percent / 100)
   const shares = (totalShares * percent) / BigInt(100)
 
   if (shares === BigInt(0)) throw new Error('Computed zero shares to withdraw; increase percent')
@@ -65,10 +57,6 @@ export const withdrawStakeByPercent = async (params: WithdrawStakeByPercentParam
   return withdrawStake({ api, operatorId, shares })
 }
 
-/**
- * Withdraw an approximate balance amount (in tokens). Computes the corresponding shares using
- * the current position and rounds down. Returns a submittable extrinsic.
- */
 export const withdrawStakeByValue = async (params: WithdrawStakeByValueParams) => {
   const { api, operatorId, account } = params
   const requestedAmount = BigInt(parseString(params.amountToWithdraw))
@@ -85,7 +73,6 @@ export const withdrawStakeByValue = async (params: WithdrawStakeByValueParams) =
   const effectiveAmount =
     requestedAmount > currentStakedValue ? currentStakedValue : requestedAmount
 
-  // shares = floor(min(requestedAmount, currentStakedValue) * totalShares / currentStakedValue)
   const shares = (effectiveAmount * totalShares) / currentStakedValue
 
   if (shares === BigInt(0))
